Create a script file
Create a script file that contains all the logic behind the game.
This file contains three main functions, computerPlay is responsible for randomly selects from the selections list, playRound function which takes in two parameters on for payerSelection and the other for computerSelections and this function is responsible for displaying the result of the game, and the last function which is game and this one is responsible for limiting the game for 5 attempts only.